% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdbwhte.R
\name{glance.rdbwhte}
\alias{glance.rdbwhte}
\title{Glance at a \code{rdbwhte} object}
\usage{
\method{glance}{rdbwhte}(x, ...)
}
\arguments{
\item{x}{A \code{rdbwhte} object.}

\item{...}{ignored.}
}
\value{
A one-row data frame.
}
\description{
broom-compatible \code{glance()} method. Returns a one-row summary of the
bandwidth-selection call.
}
\keyword{internal}
