% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdhte.R
\name{print.rdhte}
\alias{print.rdhte}
\title{Internal function.}
\usage{
\method{print}{rdhte}(x, ...)
}
\arguments{
\item{x}{Class \code{rdhte} objects.}
}
\value{
No return value, called for side effects.
}
\description{
Internal function.
}
\keyword{internal}
