% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdbwhte.R
\name{print.summary.rdbwhte}
\alias{print.summary.rdbwhte}
\title{Internal function.}
\usage{
\method{print}{summary.rdbwhte}(x, ...)
}
\arguments{
\item{x}{Class \code{summary.rdbwhte} object.}

\item{...}{ignored.}
}
\value{
The input \code{x} returned invisibly; called for side effects.
}
\description{
Internal function.
}
\keyword{internal}
