% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdhte.R
\name{print.summary.rdhte}
\alias{print.summary.rdhte}
\title{Internal function.}
\usage{
\method{print}{summary.rdhte}(x, ...)
}
\arguments{
\item{x}{Class \code{summary.rdhte} object returned by \code{summary.rdhte}.}

\item{...}{ignored.}
}
\value{
The input \code{x} returned invisibly; called for side effects.
}
\description{
Internal function.
}
\keyword{internal}
