% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdbwhte.R
\name{summary.rdbwhte}
\alias{summary.rdbwhte}
\title{Summary method for \code{rdbwhte} objects.}
\usage{
\method{summary}{rdbwhte}(object, ...)
}
\arguments{
\item{object}{Class \code{rdbwhte} object returned by \code{\link{rdbwhte}}.}

\item{...}{ignored.}
}
\value{
An S3 object inheriting from \code{rdbwhte} with class
\code{summary.rdbwhte}; printed by \code{print.summary.rdbwhte}.
}
\description{
Summary method for \code{rdbwhte} objects.
}
\keyword{internal}
