% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdhte.R
\name{summary.rdhte}
\alias{summary.rdhte}
\title{Summary method for \code{rdhte} objects.}
\usage{
\method{summary}{rdhte}(object, ...)
}
\arguments{
\item{object}{Class \code{rdhte} object returned by \code{\link{rdhte}}.}

\item{...}{ignored.}
}
\value{
An S3 object inheriting from \code{rdhte} with class
\code{summary.rdhte}; printed by \code{print.summary.rdhte}.
}
\description{
Summary method for \code{rdhte} objects.
}
\keyword{internal}
