% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdbwhte.R
\name{summary.rdbwhte}
\alias{summary.rdbwhte}
\title{Internal function.}
\usage{
\method{summary}{rdbwhte}(object, ...)
}
\arguments{
\item{object}{Class \code{rdhte} objects.}
}
\value{
No return value, called for side effects.
}
\description{
Internal function.
}
\keyword{internal}
